UHF-13337: Merged FormHooks and TitleSuggestionFormAlter
